+++ kubernetes-client-api/src/test/java/io/fabric8/kubernetes/client/extended/leaderelection/resourcelock/LeaseLockTest.java
@@ -23,6 +23,7 @@ import io.fabric8.kubernetes.client.dsl.base.PatchContext;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
+import org.mockito.Mockito;
 
 import java.time.Duration;
 import java.time.ZoneId;
@@ -82,10 +83,9 @@ class LeaseLockTest {
     lease.setMetadata(new ObjectMetaBuilder().withResourceVersion("313373").build());
     final LeaseLock lock = new LeaseLock("namespace", "name", "1337");
     // When
-    final LeaderElectionRecord result = lock.toRecordInternal(lease);
+    final LeaderElectionRecord result = lock.toRecord(lease);
     // Then
     assertNotNull(result);
-    assertEquals("313373", result.getVersion());
     assertEquals("1337", result.getHolderIdentity());
     assertEquals(15, result.getLeaseDuration().getSeconds());
     assertEquals(ZonedDateTime.of(2015, 10, 21, 4, 29, 0, 0, ZoneId.of("UTC")), result.getAcquireTime());
@@ -108,13 +108,13 @@ class LeaseLockTest {
     // Given
     final LeaderElectionRecord record = new LeaderElectionRecord(
         "1337", Duration.ofSeconds(1), ZonedDateTime.now(), ZonedDateTime.now(), 0);
-    record.setVersion("313373");
     final LeaseLock lock = new LeaseLock("namespace", "name", "1337");
     Lease lease = lock.toResource(record, lock.getObjectMeta("313373"));
+    lock.setResource(lease);
     // When
-    lock.update(kc, record);
+    lock.update(kc, record.toBuilder().leaseDuration(Duration.ofSeconds(2)).build());
     // Then
-    verify(kc.resource(lease)).patch(any(PatchContext.class));
+    verify(kc.resource(Mockito.any(Lease.class))).patch(any(PatchContext.class));
   }
 
   @Test