+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix #4963: Openshift Client return 403 when use websocket
 
 #### Improvements
+* Fix #4477 exposing LeaderElector.release to force an elector to give up the lease
 
 #### Dependency Upgrade